feat(auth-token): moved "setToken" functionality to builder
